ml_backend/requirements.txt
tensorflow
pillow
numpy
gradio